replace for-loop with foreach from stream-api
the for loop needs the length of the current string and than access the
char character by character via charAt. this prevents the JsonString
from reloading its content. the foreach from stream retrieve char by
char without knowing the end. so the jsonstring can be extended and
reloading his content from a source (like a database). this approach
also fits the buffer of the generator > the string supplies the char
at the same speed at which the generator can fill and write the buffer.

Signed-off-by: Sebastian Toepfer <sebastian+eclipse@toepfers.eu>